matching/requirements.txt
praw==5.4.0
numpy==1.18.1
protobuf==3.11.3
google-cloud==0.34.0